pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "afsp"
version = "1.0.0"
description = "Agent File Scope Protocol — filesystem perception layer for agents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "click>=8.1.0",
    "bcrypt>=4.1.0",
    "PyJWT>=2.8.0",
    "watchdog>=3.0.0",
    "PyYAML>=6.0",
    "aiosqlite>=0.19.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[project.scripts]
afsp = "afsp.cli.main:cli"

[tool.setuptools.packages.find]
include = ["afsp*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
